git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@2697 21252fe5-60e8-0310-8ff0-c1ae1001594e

+++ 3rd-party/aterm/include/_aterm.h
@@ -11,8 +11,8 @@ extern "C"
 typedef struct ProtEntry
 {
 	struct ProtEntry *next;
-	ATerm *start;
-	int    size;
+	ATerm            *start;
+	unsigned long     size;
 } ProtEntry;
 
 #define STRING_MARK 0xFF /* marker for binary strings */
@@ -22,7 +22,7 @@ extern ATbool silent;
 extern ATbool low_memory;
 
 extern ProtEntry **at_prot_table;
-extern int at_prot_table_size;
+extern unsigned int at_prot_table_size;
 extern ProtEntry *at_prot_memory;
 extern ATermProtFunc *at_prot_functions;
 extern int at_prot_functions_count;
