+++ 3rd-party/aterm/include/md5.h
@@ -24,6 +24,8 @@ These notices must be retained in any copies of any part of this
 documentation and/or software.
  */
 
+#include <stdint.h>
+
 /* GLOBAL.H - RSAREF types and constants
  */
 
@@ -46,10 +48,10 @@ extern "C"
 typedef unsigned char *POINTER;
 
 /* UINT2 defines a two byte word */
-typedef unsigned short int UINT2;
+typedef uint16_t UINT2;
 
 /* UINT4 defines a four byte word */
-typedef unsigned long int UINT4;
+typedef uint32_t UINT4;
 
 /* PROTO_LIST is defined depending on how PROTOTYPES is defined above.
 If using PROTOTYPES, then PROTO_LIST returns the list, otherwise it
